Update patch set 1
Patch Set 1:

3.16.0 is the upper bound.

The lower bound is wrong, per the bug I linked to.  1.14.0 will not actually work correctly.  But it's too late to change the minimum requirements upstream for stable/newton.

Comment #5 in the bug I linked indicates that the safest route would be to pick the versions from upper-constraints, which is probably true, but to my knowledge we aren't currently following that strategy.

Patch-set: 1
Reviewer: Gerrit User 75 <75@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0