[WIP] Seperate entrypoint for tempest plugins
Previously python-cinder package installed the entry point for the in-tree
Tempest plugin which cause failure while running tempest list-plugins
because Tempest found the entry point but does not have Tempest plugin
code which is subpackaged in python-cinder-tests.
Separate cinder_tests egg-info is created with the plugin entry point
so that Tempest includes the plugin only when subpackage is installed.

Change-Id: Ic815cc666bcefcb512af026f3106a4977aa879ea