Update patch set 2
Patch Set 2: Code-Review+2

All WeiRDO jobs are failing because this review bumps the requirement for oslo.versionedobjects to >= 1.19.0, and the highest version that passed CI tests was 1.18.0.

This should not be an issue for this package, as it will be tested for promotion with the current version (1.20.0).

Patch-set: 2
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+2