Update patch set 4
Patch Set 4:

Build is passing with concurrency to 1 but it's taking 24 minutes to build it (14 mins running unit tests). I wonder if it's doable for the  DLRN instance which is building on each commit. AFAIK, there is no way to run a reduced set of unit tests that in a meaningful way (correct me if i wrong), so i guess we can move on with this and observe the impact in the dlrn instance.

Patch-set: 4
Label: Code-Review=0