Update patch set 1
Patch Set 1:

testr process running unit tests was hang for more that a day and i killed it. Process seemed to be reading in stdin, but it sounds a bit weird to me. Let's see if it works fine in next builds.

Patch-set: 1
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0