Add resource_filters.json file to config dir
The relies on the resource_filters.json file which
is loaded on startup and contains the allowed query
options for API calls.

When this file does not exist all API calls to endpoints
trying to use query options for filtering will fail with
a 400 bad request because it's not allowed.

After downloading this file [1] and placing it in the correct
place [2] and restart the service the API calls were working.

[1] https://github.com/openstack/cinder/blob/stable/rocky/etc/cinder/resource_filters.json
[2] https://github.com/openstack/cinder/blob/4bc1e5f6a33f66945b5a01e6d5b3d3688ae647d5/cinder/api/common.py#L40

Change-Id: Ie439dc99da367d06c54948f674c462acf0c43c06