Create python-cinder-common subpackage
This patch splits the current python-cinder package in 2

- python-cinder
- python-cinder-common

And makes python-cinder depend on python-cinder-common which contains
the Cinder code and the minimum dependencies required to use Cinder
drivers without running the Cinder services.

This allows other packages, like python-cinderlib, reduce their total
dependencies by depending on the python-cinder-common package and
leveraging the mocking feature in cinderlib added by patch
https://review.opendev.org/655401.

This is the equivalent of the master branch patch, but with support for
Python 2 and Python 3.

Co-Authored-By: Gorka Eguileor <geguileo@redhat.com>
Change-Id: Id95890090d8436e026369df8eb913726fa8ddbc0