WIP: Create python-cinder-common subpackage
This patch splits the current python-cinder package in 2

- python-cinder
- python-cinder-common

And makes python-cinder depend on python-cinder-common which contains
the Cinder code and the minimum dependencies required to use Cinder
drivers without running the Cinder services.

This allows other packages, like python-cinderlib, reduce their total
dependencies.

This will require changes to the cinderlib package to make it depend on
python-cinder-common as well as path https://review.opendev.org/655401
to allow cinderlib to work without those dependencies.

Change-Id: Id95890090d8436e026369df8eb913726fa8ddbc0