Update patch set 1
Patch Set 1: Code-Review-1

In RDO, we use python-lxml-3.2.1 provided in CentOS/RHEL base repo. The recommended practice is not to provide updated packages in RDO for the ones in base repos as we could break others, unless it's justified with known issues. Is it the case for cinder and lxml? is this bump required for any specific issue we will hit?

Patch-set: 1
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=-1