Update patch set 1
Patch Set 1: Code-Review-1

legacy-DLRN-build-tripleo is failing with:-
 DEBUG: BUILDSTDERR:            Requires: python2-os-brick >= 2.8.0
2019-03-12 18:14:00.402092 | rdo-centos-7 | DEBUG: BUILDSTDERR:            Available: python2-os-brick-2.7.0-0.20190124090828.d7ebf97.el7.noarch (delorean-current-tripleo)
2019-03-12 18:14:00.402164 | rdo-centos-7 | DEBUG: BUILDSTDERR:                python2-os-brick = 2.7.0-0.20190124090828.d7ebf97.el7

This is because os-brick 2.8.0 is not there in promoted current-tripleo repo:- https://trunk.rdoproject.org/centos7-master/current-tripleo/ , this repo is used in tripleo check jobs. So if cinder package is built somehow directly in cinder gate or via depends-on it will fail. So i think good to hold this until next promotion.

Patch-set: 1