Update patch set 2
Patch Set 2:

Yes, the reason I added the requirement here instead of tooz is; a) not everyone who wants tooz needs the etcd driver, and b) it's TripleO's cinder that specifically selected etcd for its DLM.

I've been working with John Eckersberg to get the package into f28's updates repo, and it's currently "baking" in updates-testing. I see that Alredo added more karma to help it along, but I was going to ask John to do a force, which I gather is acceptable after a week or so in testing.

Anyway, once it's in f28 updates then I plan to recheck, and see if legacy-DLRN-rpmbuild-fedora passes.

Patch-set: 2