Create /var/lib/cinder/groups directory
When using the default configs for the
Tooz coordinator it uses the file backend.

The file backend default to using the
$state_path set in cinder.conf which by
default is /var/lib/cinder so it searches
for the /var/lib/cinder/groups directory
which doesn't exist.

Without this directory the API service
doesn't start.

Requires backport to Train as well where
I observed this behavior.

Change-Id: If747a84fc6a270673f5def542392793d11d4e8a8