Update patch set 1
Patch Set 1: Code-Review-1

What's the error that you get?

Because the driver tries to create the required directories on start [1][2]. 
And if that's not working as it should, then we either have permission problems that prevent us from creating the directory, or there is a bug in TooZ or Cinder is somehow using Tooz incorrectly.

I would like to know what's going on before we merge this, because afaik this should work just fine without us creating that directory.

[1]: https://github.com/openstack/tooz/blob/7147939860068b832a4cca0ece564a35da09bc57/tooz/drivers/file.py#L278
[2]: https://github.com/openstack/tooz/blob/master/tooz/drivers/file.py#L237

Patch-set: 1
Reviewer: Gerrit User 181 <181@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=-1