Remove version requirement from python3-google-api-client
[1] set a minimum requirement for python3-google-api-client of
version 1.11.0. However, the latest version we have in the RDO
requirements is 1.6.7 [2]. The version bump passed CI only because
the RDO package has Epoch: 1.

Since the package situation in Fedora is currently complicated [3], this
commit removes the minimum version requirement to avoid confusion.
If a version bump is really required, it will take some time.

[1] - https://review.rdoproject.org/r/31895
[2] - https://cbs.centos.org/koji/packageinfo?packageID=2346
[3] - https://src.fedoraproject.org/rpms/google-api-python-client/commits/rawhide

Change-Id: Ie413235071336ae82f5fe11718acdfff4318a7bb