Drop the capping for python3-google-api-client
Last version in CentOS 8 HA repo is 1.6.5 and RDO provides 1.6.7.
Currently we can not provide 1.11.0 as it's not in Fedora and there are
some issues to updated it, and in Cinder is only used in backup GCS
driver which is optional and actually not tested so i think we can
temporarily live with the version in CentOS repo.

Note that bumping the minimal version to 1.11.0 didn't break RDO just
because the build in RDO has epoch 1 but that's not correct, so let's
fix it.

Change-Id: Ib0c028d0d7e61f9de86c7b3bb7fd4bbd23683276