Fix the dependency on libcgroup-tools (no arch)
I've originally added the arch specifier while trying to override
the bogus lintian warning about avoiding direct dependency on libraries
(but the libcgroup-tools contains programs), and following some
guidelines for those cases I've added the arch.

But it doesn't make sense in this context, and it makes arch-aware
a noarch package.

Change-Id: I9fb20e0ba2622045fa38d9901b00b7fb65f0351a