Update patch set 1
Patch Set 1:

Luigi, I'm not very familiar with this kind of case, but from a theoretical point of view this seems like something that should be changed in the downstream spec instead of upstream.  Why are we changing it upstream instead?  Is it to remove merge conflicts on future changes?

Patch-set: 1
Reviewer: Gerrit User 181 <181@270e2033-b340-4cff-9539-693957ebf0e7>