Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Luigi, I'm not very familiar with this kind of case, but from a theoretical point of view this seems like something that should be changed in the downstream spec instead of upstream.  Why are we changing it upstream instead?  Is it to remove merge conflicts on future changes?
> 
> Jon and Lon can comment on this, but I think the idea is that the packaging is done here and the delta should be none. We do it already for other components:
> https://review.rdoproject.org/r/c/openstack/tempest-distgit/+/37006

Correct the guard change here is a no-op for RDO packaging but when we build for downstream the variable for RHOSP is changed and the packaging change is then in effect.  This is to allow OSP to not have to ship extra libraries we don't support or use.

Patch-set: 1
Reviewer: Gerrit User 127 <127@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1