Add requires python3-oslo-policy >= 4.3.0
When SLURP upgrading from Antelope to Caracal the
Secure RBAC is completed which means that
enforce_scope is removed and the logic is enabled
default.

If oslo.policy is not upgraded when moving from
Antelope to Caracal the enforce_scope=False will
still be the default, so even if an operator
has fixed the service roles they must know to
upgrade oslo.policy specifically to get it
working.

Change-Id: Id593cd7bed157800cba03d7e6271b8a6744412b4